+++ python-libraries/data-platform-catalogue/CHANGELOG.md
@@ -7,6 +7,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.10.0] 2024-02-01
+
+### Changed
+
+- Custom properties are now added to the metadata of each search result
+- Datasets return domain information
+- Domain information is now returned as `domain_id` and `domain_name` metadata
+
 ## [0.9.0] 2024-01-31
 
 ### Added